doubleとして表示できていなかったのを修正。

+++ ZipLengthTester/ZipFileOperation.cs
@@ -22,7 +22,7 @@ namespace ZipLengthTester
             {
                 var totalLength = za.Entries.Sum(entry => entry.Length);
                 var kiro = thousandIsBit ? 1024 : 1000;
-                d = totalLength / (kiro ^ multiplier);
+                d = (double) totalLength / (double) (kiro ^ multiplier);
 
             }
